Use pbfparser instead of XML parser in map-creator
