+++ brouter-map-creator/build.gradle
@@ -2,11 +2,12 @@ plugins {
     id 'java-library'
 }
 
-
 dependencies {
     implementation project(':brouter-codec')
     implementation project(':brouter-util')
     implementation project(':brouter-expressions')
-	
+
+    implementation group: 'org.openstreetmap.osmosis', name: 'osmosis-osm-binary', version: '0.48.3'
+
     testImplementation('junit:junit:4.13.1')
 }
