+++ brouter-map-creator/src/main/java/btools/mapcreator/BPbfFieldDecoder.java
